update calculator and catalog
