+++ internal/telegram/handler_menu.go
@@ -4,11 +4,8 @@ import (
 	"context"
 	"errors"
 	"fmt"
-
 	tg "github.com/go-telegram-bot-api/telegram-bot-api/v5"
 	"github.com/sonyamoonglade/poison-tg/internal/domain"
-	"github.com/sonyamoonglade/poison-tg/internal/repositories/dto"
-	"github.com/sonyamoonglade/poison-tg/pkg/functools"
 )
 
 func (h *handler) Start(ctx context.Context, m *tg.Message) error {
@@ -31,7 +28,17 @@ func (h *handler) Start(ctx context.Context, m *tg.Message) error {
 			return err
 		}
 	}
-	return h.sendWithKeyboard(chatID, getStartTemplate(username), initialMenuKeyboard)
+
+	if err := h.sendWithKeyboard(chatID, getStartTemplate(username), initialMenuKeyboard); err != nil {
+		return err
+	}
+
+	yuanRate, err := h.yuanService.GetRate()
+	if err != nil {
+		return err
+	}
+
+	return h.cleanSend(tg.NewMessage(chatID, fmt.Sprintf("Курс юаня на сегодня: %.2f ₽", yuanRate)))
 }
 
 func (h *handler) Menu(ctx context.Context, chatID int64) error {
@@ -39,90 +46,6 @@ func (h *handler) Menu(ctx context.Context, chatID int64) error {
 		return err
 	}
 	return h.sendWithKeyboard(chatID, getTemplate().Menu, menuButtons)
-}
-
-func (h *handler) Catalog(ctx context.Context, chatID int64) error {
-	var (
-		telegramID = chatID
-		first      bool
-	)
-
-	customer, err := h.customerRepo.GetByTelegramID(ctx, telegramID)
-	if err != nil {
-		return err
-	}
-
-	if err := h.cleanSend(tg.NewMessage(chatID, getCatalog(*customer.Username))); err != nil {
-		return err
-	}
-
-	// Load appropriate item
-	item := h.catalogProvider.LoadAt(customer.CatalogOffset)
-
-	thumbnails := functools.Map(func(url string) interface{} {
-		thumbnail := tg.NewInputMediaPhoto(tg.FileURL(url))
-		if !first {
-			// add caption to first element
-			thumbnail.Caption = item.GetCaption()
-			first = true
-		}
-		return thumbnail
-	}, item.ImageURLs)
-
-	group := tg.NewMediaGroup(chatID, thumbnails)
-
-	// Sends thumnails with caption
-	sentMsgs, err := h.b.client.SendMediaGroup(group)
-	if err != nil {
-		return err
-	}
-
-	msgIDs := functools.Map(func(m tg.Message) int {
-		return m.MessageID
-	}, sentMsgs)
-
-	// Prepare buttons for controlling prev, next
-	var (
-		currentOffset = customer.CatalogOffset
-		hasNext       = h.catalogProvider.HasNext(currentOffset)
-		hasPrev       = h.catalogProvider.HasPrev(currentOffset)
-	)
-
-	btnArgs := catalogButtonsArgs{
-		hasNext: hasNext,
-		hasPrev: hasPrev,
-		msgIDs:  msgIDs,
-	}
-
-	if hasNext {
-		next := h.catalogProvider.LoadNext(currentOffset)
-		btnArgs.nextTitle = next.Title
-	} else if hasPrev {
-		prev := h.catalogProvider.LoadPrev(currentOffset)
-		btnArgs.prevTitle = prev.Title
-	}
-
-	buttons := prepareCatalogButtons(btnArgs)
-	return h.sendWithKeyboard(chatID, "Кнопки для пролистывания каталога", buttons)
-}
-
-func (h *handler) Calculator(ctx context.Context, chatID int64) error {
-	var telegramID = chatID
-
-	customer, err := h.customerRepo.GetByTelegramID(ctx, telegramID)
-	if err != nil {
-		return fmt.Errorf("customerRepo.GetByTelegramID: %w", err)
-	}
-
-	updateDTO := dto.UpdateCustomerDTO{
-		State: &domain.StateWaitingForCalculatorInput,
-	}
-
-	if err := h.customerRepo.Update(ctx, customer.CustomerID, updateDTO); err != nil {
-		return fmt.Errorf("customerRepo.Update: %w", err)
-	}
-
-	return h.cleanSend(tg.NewMessage(chatID, "Отправь мне цену товара в юанях, а я скажу сколько это будет стоить в переводе на рубли для тебя"))
 }
 
 func (h *handler) MyOrders(ctx context.Context, chatID int64) error {
