+++ templates.json
@@ -18,5 +18,5 @@
   "afterPaid": "%s, твой заказ %s сейчас на подтверждении у админа. Он напишет тебе в личные сообщения и подтвердит статус покупки.\n\nНе нужно отправлять деньги никому кроме бота! Даже админу!\n\nТолько админ проверяет поступление денег и обозначает статус покупки. Он указан в шапке группы",
   "myOrdersStart":"Вот твои заказы, %s!\n\n",
   "myOrdersEnd": "-----------\n\n",
-  "singleOrderPreview": "Заказ: %s\nОплачен: %s\nПодтвержден админом: %s\nТип: %s\nАдрес доставки: %s\nТоваров в корзине: %d\nСумма в юанях: %d\nСумма в рублях: %d\n\nТовары:\n"
+  "singleOrderPreview": "Заказ: %s\nОплачен: %s\nПодтвержден админом: %s\nТип: %s\nАдрес доставки: %s\nТоваров в корзине: %d\nСумма в юанях: %d ¥\nСумма в рублях: %d ₽\n\nТовары:\n"
 }
